Add BlackberryNDK script for ndk related operations
Changelog:
- Added UnitTest context manager class to be used to simplified unit tests implementation
- Added BlackberryNDK script to encapsulate interactions with the ndk tools
- Start using the blackberryNDK class in project.py and blackberry.py
- added .swp extension to the ignore list of the sdk packager
- add importing of a new project in a workspace so it can be built
- refactored builder.py a bit
- add execute flag on builder.py and blackberry.py

NOTE: The paths for windows will need to be validated and updated if necessary

NOTE: the project name must match in the .project file and in the folder name for the import and build steps to work.  (This is an issue for as we copy the HelloWorldDisplay sample, but will be resolved once we create our own project.)
There must not be any spaces in the project name or path to the project for the build to succeed.  Will have to work that issue out with the studio team and possibly RIM as it's a shortcoming of their tools.

TESTS:
All BlackberryNDK unit tests are successful

Built and installed mobilesdk and successfully created a new project with the name HelloWorldDisplay and then successfully build and run it on the simulator using the project and builder scripts from command line

./project.py HelloWorldDisplay com.macadamian.helloworlddisplay ~/src/tsw blackberry

blackberry/builder.py run -t simulator -d ~/src/tsw/HelloWorldDisplay/build/blackberry/HelloWorldDisplay -p someNdk