TIMOB-9665 Blackberry: fix property enumeration on JS object created by the framework
Reviewer: DC

The lookup of inherited methods such as toString() was broken for Ti instances that we create, resulting in a crash.
The cause is that by returning undefined in the lookup the V8 internal lookup was not invoked.

Changelog:
- Return an empty value handle when not finding a keyword in our framework
- Protect the TiUIBase initializeTiObject so that it can be called from derived classes without negative effects.
- Reverted the calls from derived classes to call in their direct parent

Tests:
Built for simulator, verified that passing a Ti object to a function that expects a string, ends up invoking the toString on the object and it doesn't crash the JS anymore.
=> verified that the event callbacks are not broken by the change
=> Using breakpoints, verified that the prop_setter function is still getting called with the change